fix: separando funcoes por arquivos
